tests/wpt/tests/css/css-scroll-snap/scroll-initial-target/parsing/scroll-initial-target-computed.html

<!DOCTYPE html>
<html>

<head>
  <meta charset="utf-8">
  <title> CSS Scroll Snap 2 Test: scroll-initial-target-* computed values</title>
  <link rel="help" href="https://drafts.csswg.org/css-scroll-snap-2/#scroll-initial-target">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/computed-testcommon.js"></script>
</head>

<body>
  <div id="target"></div>
  <script>
    test_computed_value("scroll-initial-target", "nearest");
    test_computed_value("scroll-initial-target", "none");
    </script>
</body>

</html>
